Move CLI out of __init__
